[PM] GetAvailableFrequencies vector type changed
From vector of 'unsigned long' to vector of 'uint32_t'.

+++ bbque/pm/power_manager.cc
@@ -208,7 +208,7 @@ PowerManager::GetClockFrequencyInfo(
 PowerManager::PMResult
 PowerManager::GetAvailableFrequencies(
 		ResourcePathPtr_t const & rp,
-		std::vector<unsigned long> & freqs) {
+		std::vector<uint32_t> & freqs) {
 	switch (rp->ParentType(rp->Type())) {
 	case br::ResourceIdentifier::GPU:
 		if (!gpu) break;
